requirement.txt
Flask
mysql-connector-python
flask_mysqldb
pillow